Exercise20_23
Modify LiveExample 20.9 EvaluateExpression.java to add operators ^ for exponent and % for remainder. For example, 3 ^ 2 is 9 and 3 % 2 is 1. The ^ operator has the highest precedence and the % operator has the same precedence as the * and / operators. For simplicity, assume the expression has at most one ^ operator.

Your program should prompt the user to enter an expression, evaluate the expression, and display the result.