Exercise28_05.java
Define a new class named UnweightedGraphWithGetPath that extends UnweightedGraph with a new method for finding a path between two vertices with the following header:

public List<Integer> getPath(int u, int v);

The method returns a List<Integer> that contains all the vertices in a path from u to v in this order. Using the BFS approach, you can obtain the shortest path from u to v. If there isn’t a path from u to v, the method returns null. Write a test program that creates a graph for Figure 28.1. The program prompts the user to enter two cities and displays their paths.